Rename module folder and upgrade poetry and pyproject.toml.
