+++ pyproject.toml
@@ -1,10 +1,17 @@
-[tool.poetry]
-name = "basic-strategy"
+[project]
+name = "blackjack-trainer"
 version = "0.1.0"
 description = ""
-authors = ["jonas.schrage <119843859+joschrag@users.noreply.github.com>"]
+authors = [{name="jonas.schrage", email="<119843859+joschrag@users.noreply.github.com>"}]
+keywords = ["python","dash","blackjack","basic","strategy"]
 license = "MIT"
 readme = "README.md"
+requires-python = ">=3.9"
+dynamic = ["dependencies"]
+
+[project.urls]
+repository = "https://github.com/joschrag/blackjack_trainer"
+
 
 [tool.poetry.dependencies]
 python = "^3.12"
@@ -12,13 +19,13 @@ pandas = "^2.2.2"
 matplotlib = "^3.9.0"
 plotly = "^5.22.0"
 sqlalchemy = "^2.0.30"
-dash = {extras = ["diskcache"], version = "^2.17.1"}
-dash-bootstrap-components = "^1.6.0"
 pathlib = "^1.0.1"
 urllib3 = "^1.9.0"
 dash-mantine-components = "0.13.0a2"
 dash-iconify = "^0.1.2"
 pyyaml = "^6.0.1"
+dash = {extras = ["diskcache", "testing"], version = "^3.0.4"}
+dash-bootstrap-components = "^2.0.3"
 
 
 [tool.poetry.group.dev.dependencies]
@@ -31,7 +38,6 @@ ipykernel = "^6.29.4"
 flake8 = "^7.0.0"
 pre-commit = "^3.7.1"
 pytest-cov = "^5.0.0"
-dash = {extras = ["testing"], version = "^2.17.1"}
 pydocstyle = "^6.3.0"
 pytest-mock = "^3.14.0"
 types-pyyaml = "^6.0.12.20240311"
@@ -45,3 +51,9 @@ line-length = 119
 
 [tool.isort]
 profile="black"
+
+[tool.pytest.ini_options]
+addopts = "-ra -q --webdriver Firefox"
+testpaths = [
+    "tests",
+]
