+++ src/blackjack_trainer/app/pages/settings.py
