+++ src/blackjack_trainer/blackjack/hand.py
